+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix "logboog" typo in preview message
 - Improve minimal UI switching
+- Don't reset the elapsed time to 0 when pausing a recording
 
 ## 0.6.0